server/requirements.txt
asgiref==3.8.1
boto3==1.35.85
botocore==1.35.85
certifi==2024.12.14
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
